% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{convert_json_to_df}
\alias{convert_json_to_df}
\title{Convert json to dataframe for xyz R and l and euler angles with timestamp.
useful for VR data recorded with the vectorMatch game}
\usage{
convert_json_to_df(json_object)
}
\arguments{
\item{json_object}{structure as returned from jsonlite::read_json}
}
\value{
res dataframe that contains right_position_x, timestamp, etc.
}
\description{
Convert json to dataframe for xyz R and l and euler angles with timestamp.
useful for VR data recorded with the vectorMatch game
}
