% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{dcrb}
\alias{dcrb}
\title{Do Call Rbind on list of dataframes}
\usage{
dcrb(list_of_dataframes)
}
\arguments{
\item{list}{of dataframes}
}
\value{
df row-bound concatenated dataframe
}
\description{
Do Call Rbind on list of dataframes
}
